+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to the "crates" extension will be documented in this file.
 ### 0.2.1
 * Patreon link changed to a badge.
 * Marketplace badges are linked to marketplace. [#13](https://github.com/serayuzgur/crates/issues/13).
+* Package json improved.
+* The up to date decorator (👍) is now customizable from setting crates.upToDateDecorator [#12](https://github.com/serayuzgur/crates/issues/12).
 
 
 ### 0.2.0